pyproject.toml
[project]
name = "mt5-data-pipeline"
version = "0.1.0"
description = "Data pipeline for automated trading system. Ingests OHLC data from IBKR API (EUREX/XETRA), stores in PostgreSQL/TimescaleDB."
authors = [
    { name = "Johan", email = "55276630+Hiddentale@users.noreply.github.com" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ib_async>=0.3.0",
    "psycopg2-binary>=2.9.0",
    "polars>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
